fix/integer to INT64 and float to FLOAT64

+++ local_csv_data/tabel_definitions_DDL/ddl_person.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE testRoxDB.Person
 (
-  BusinessEntityID INT NOT NULL,
+  BusinessEntityID INT64 NOT NULL,
   PersonType STRING,
   NameStyle INT64,
   Title STRING,
